// api/lib/types.ts
export type WritingPart = 1 | 2 | 3;
// ...
export interface WritingAnswer {
  questionId: string;
  questionType: WritingPart;
  questionText: string; // User input question text
  text: string;
  wordCount: number;
  savedAt?: Date;
}